# Backend/schemas/openapi.yaml
openapi: 3.0.0
info:
  title: Book Database Api
  description: A simple API which provides access to a database of books and their authors.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://turboexhibit-diegosalsa-3030.codio-box.uk/api/v1'
    description: Development server on Codio
components:
  securitySchemes:
    jwtAuth:     
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    NotFoundError:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: JWT is missing or invalid. Please log in using the /login endpoint. 
      content:
        application/json:
          schema:
            type: string
      required:
        - message
    SchemaValidationError:
      description: Invalid resource schema provided in request body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PermissionError:
      description: Access level not sufficient for accessing requested resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerSideError:
      description: Something went wrong on the server side. If this keeps repeating, contact the admin.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
  parameters:
    page:
      in: query
      name: page
      schema:
        type: integer
        default: 0
      description: The number of the page of items to show. Each page is the size of "limit".
    limit:
      in: query
      name: limit
      schema:
        type: integer
        default: 5
      description: The numbers of items to return.
    order:
      in: query
      name: order
      schema:
        type: string
        default: "id"
      description: The property by which to sort the results.
...
paths:
...
  /user:
    description: This endpoint represents the user account resource. Each user is identified by an ID.
    get:
      security:
        - jwtAuth: []
      tags:
        - User
      description:
        Get data of all users from the database.
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/order'
      responses:
        '200':
          description: Array of all user records.
          headers:
            X-Total-Count:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: ./user.json#/definitions/userGet
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
    post:
      tags:
        - User
      description:
        Add data of a single user to the database (register).
      responses:
        '201':
          description: User account created.
        '400':
          $ref: '#/components/responses/SchemaValidationError'
        '500':
          $ref: '#/components/responses/ServerSideError'
      requestBody:
        description: All new user data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./user.json#/definitions/userAdd
  /user/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the user
    get:
      security:
        - jwtAuth: []
      tags:
        - User
      description:
        Get data of a single user.
      responses:
        '200':
          description: View of user data.
          content:
            application/json:
              schema:
                $ref: ./user.json#/definitions/userGet
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      security:
        - jwtAuth: []
      tags:
        - User
      description:
        Modify data of a single user.
      responses:
        '204':
          description: Modified user account.
        '400':
          $ref: '#/components/responses/SchemaValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerSideError'
      requestBody:
        description: Modifiable user data and user ID.
        required: true
        content:
          application/json:
            schema:
              $ref: ./user.json#/definitions/userUpd
    delete:
      security:
        - jwtAuth: []
      tags:
        - User
      description:
        Delete the data of a single user.
      responses:
        '204':
          description: Deleted user account.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerSideError'
...